requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
websockets==11.0.3
python-dotenv==1.0.0
firebase-admin==6.4.0
openai==2.2.0
aiohttp==3.13.0
aiodns==3.0.0
websocket-client==1.9.0
bcrypt==5.0.0
email-validator==2.0.0
beautifulsoup4==4.13.3
lxml==5.3.1
PyPDF2==3.0.1
pdfplumber==0.11.4
pytest==8.3.2
google-auth==2.41.1
google-auth-oauthlib==1.1.0
cryptography==46.0.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
matplotlib==3.7.1
jsonschema>=4.17.0

# Machine Learning dependencies
numpy>=1.24.0,<2.0.0
torch==2.8.0
scikit-learn==1.6.1
joblib>=1.3.0
librosa==0.11.0
soundfile>=0.12.0
speechbrain==1.0.3
